LayoutTests/fast/dom/assign-to-prototype-accessor-on-prototype-should-be-silent.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/js-test-pre.js"></script>
<script>
description("Makes sure that assigning to a DOM property directly on the prototype produces a silent error.");

Document.prototype.oncopy = "fail";
Document.prototype.selectedStylesheetSet = "fail";
Document.prototype.dir = "fail";
HTMLElement.prototype.innerText = "fail";

</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
