LayoutTests/fast/dom/assign-to-prototype-accessor-on-prototype-should-be-silent-expected.txt

CONSOLE MESSAGE: line 8: Deprecated attempt to set property 'oncopy' on a non-Document object.
CONSOLE MESSAGE: line 9: Deprecated attempt to set property 'selectedStylesheetSet' on a non-Document object.
CONSOLE MESSAGE: line 11: Deprecated attempt to set property 'innerText' on a non-HTMLElement object.
Makes sure that assigning to a DOM property directly on the prototype produces a silent error.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
